--- backend/alembic/versions/647bd0d25e3d_add_chat_tables.py
@@ -1,48 +1,0 @@
-"""add chat tables
-
-Revision ID: 647bd0d25e3d
-Revises: 52a6636186a0
-Create Date: 2026-07-02 16:31:20.002138
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '647bd0d25e3d'
-down_revision: Union[str, None] = '52a6636186a0'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('chat_conversations',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('user_id', sa.UUID(), nullable=False),
-    sa.Column('title', sa.String(length=255), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('updated_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('chat_messages',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('conversation_id', sa.UUID(), nullable=False),
-    sa.Column('role', sa.String(length=15), nullable=False),
-    sa.Column('content', sa.Text(), nullable=False),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.ForeignKeyConstraint(['conversation_id'], ['chat_conversations.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('chat_messages')
-    op.drop_table('chat_conversations')
-    # ### end Alembic commands ###
