--- backend/alembic/versions/b2c3d4e5f6a7_make_hashed_password_nullable.py
@@ -1,23 +1,0 @@
-"""make hashed_password nullable for Cognito users
-
-Revision ID: b2c3d4e5f6a7
-Revises: a1b2c3d4e5f6
-Create Date: 2026-06-22 11:30:00.000000
-
-"""
-from typing import Sequence, Union
-from alembic import op
-import sqlalchemy as sa
-
-revision: str = 'b2c3d4e5f6a7'
-down_revision: Union[str, None] = 'a1b2c3d4e5f6'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.alter_column('users', 'hashed_password', nullable=True)
-
-
-def downgrade() -> None:
-    op.alter_column('users', 'hashed_password', nullable=False)
