--- backend/alembic/versions/c9f1a2b3d4e5_add_portfolio_snapshots.py
@@ -1,38 +1,0 @@
-"""add portfolio snapshots table
-
-Revision ID: c9f1a2b3d4e5
-Revises: 52a6636186a0
-Create Date: 2026-07-15 10:00:00.000000
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-revision: str = "c9f1a2b3d4e5"
-down_revision: Union[str, None] = "647bd0d25e3d"
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    op.create_table(
-        "portfolio_snapshots",
-        sa.Column("id", sa.UUID(), server_default=sa.text("uuid_generate_v4()"), nullable=False),
-        sa.Column("portfolio_id", sa.UUID(), nullable=False),
-        sa.Column("snapshot_date", sa.Date(), nullable=False),
-        sa.Column("total_value", sa.Numeric(precision=18, scale=2), nullable=False),
-        sa.Column("benchmark_value", sa.Numeric(precision=18, scale=2), nullable=True),
-        sa.Column("created_at", sa.DateTime(timezone=True), server_default=sa.text("NOW()"), nullable=False),
-        sa.ForeignKeyConstraint(["portfolio_id"], ["portfolios.id"], ondelete="CASCADE"),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("portfolio_id", "snapshot_date", name="uq_portfolio_snapshot_date"),
-    )
-    op.create_index("ix_portfolio_snapshots_portfolio_id", "portfolio_snapshots", ["portfolio_id"])
-
-
-def downgrade() -> None:
-    op.drop_index("ix_portfolio_snapshots_portfolio_id", table_name="portfolio_snapshots")
-    op.drop_table("portfolio_snapshots")
